8255001: Move G1PeriodicGCTask to its own file
