+++ src/hotspot/share/gc/g1/g1PeriodicGCTask.hpp
@@ -1,0 +1,42 @@
+/*
+ * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ *
+ */
+
+#ifndef SHARE_GC_G1_G1PERIODICGCTASK_HPP
+#define SHARE_GC_G1_G1PERIODICGCTASK_HPP
+
+#include "gc/g1/g1ServiceThread.hpp"
+#include "utilities/globalDefinitions.hpp"
+#include "utilities/ticks.hpp"
+
+// Task handling periodic GCs
+class G1PeriodicGCTask : public G1ServiceTask {
+  bool should_start_periodic_gc();
+  void check_for_periodic_gc();
+
+public:
+  G1PeriodicGCTask(const char* name);
+  virtual void execute();
+};
+
+#endif // SHARE_GC_G1_G1PERIODICGCTASK_HPP
